Replace Time Shift dropdown with new dropdown
